+++ src/main/java/com/arobs/weather/entity/WeatherLocation.java
@@ -1,77 +1,77 @@
 package com.arobs.weather.entity;
 
+import org.jdto.annotation.Source;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
-import org.jdto.annotation.Source;
-
 @Entity
 @Table(name = "weather_location")
 public class WeatherLocation {
-	private Integer id;
-	private String name;
-	private String countryCode;
-	private String countyCode;
-	private Double lon;
-	private Double lat;
+    private Long id;
+    private String name;
+    private String countryCode;
+    private String countyCode;
+    private Double lon;
+    private Double lat;
 
-	@Id
-	@Column(name = "id")
-	public Integer getId() {
-		return id;
-	}
+    @Id
+    @Column(name = "id")
+    public Long getId() {
+        return id;
+    }
 
-	public void setId(Integer id) {
-		this.id = id;
-	}
+    public void setId(Long id) {
+        this.id = id;
+    }
 
-	@Column(name = "name")
-	public String getName() {
-		return name;
-	}
+    @Column(name = "name")
+    public String getName() {
+        return name;
+    }
 
-	public void setName(String name) {
-		this.name = name;
-	}
+    public void setName(String name) {
+        this.name = name;
+    }
 
-	@Column(name = "country")
-	@Source("country")
-	public String getCountryCode() {
-		return countryCode;
-	}
+    @Column(name = "country")
+    @Source("country")
+    public String getCountryCode() {
+        return countryCode;
+    }
 
-	public void setCountryCode(String countryCode) {
-		this.countryCode = countryCode;
-	}
+    public void setCountryCode(String countryCode) {
+        this.countryCode = countryCode;
+    }
 
-	@Column(name = "county")
-	public String getCountyCode() {
-		return countyCode;
-	}
+    @Column(name = "county")
+    public String getCountyCode() {
+        return countyCode;
+    }
 
-	public void setCountyCode(String countyCode) {
-		this.countyCode = countyCode;
-	}
+    public void setCountyCode(String countyCode) {
+        this.countyCode = countyCode;
+    }
 
-	@Column(name = "lon")
-	@Source( "coord.lon")
-	public Double getLon() {
-		return lon;
-	}
+    @Column(name = "lon")
+    @Source("coord.lon")
+    public Double getLon() {
+        return lon;
+    }
 
-	public void setLon(Double lon) {
-		this.lon = lon;
-	}
+    public void setLon(Double lon) {
+        this.lon = lon;
+    }
 
-	@Column(name = "lat")
-	@Source( "coord.lat")
-	public Double getLat() {
-		return lat;
-	}
+    @Column(name = "lat")
+    @Source("coord.lat")
+    public Double getLat() {
+        return lat;
+    }
 
-	public void setLat(Double lat) {
-		this.lat = lat;
-	}
+    public void setLat(Double lat) {
+        this.lat = lat;
+    }
 }
